+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,85 @@
 # @agentflow/cli
 
+## 0.2.8
+
+### Patch Changes
+
+- Bymp
+- Updated dependencies
+  - create-agentflow@0.1.9
+  - @agentflow/core@0.2.8
+  - @agentflow/tools@0.1.9
+
+## 0.2.7
+
+### Patch Changes
+
+- Force bump
+- Updated dependencies
+  - @agentflow/core@0.2.7
+  - create-agentflow@0.1.8
+  - @agentflow/tools@0.1.8
+
+## 0.2.6
+
+### Patch Changes
+
+- Force bump
+- Updated dependencies
+  - @agentflow/core@0.2.6
+  - create-agentflow@0.1.7
+  - @agentflow/tools@0.1.7
+
+## 0.2.5
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @agentflow/tools@0.1.6
+  - @agentflow/core@0.2.5
+  - create-agentflow@0.1.6
+
+## 0.2.4
+
+### Patch Changes
+
+- Force
+- Updated dependencies
+  - @agentflow/tools@0.1.5
+  - @agentflow/core@0.2.4
+  - create-agentflow@0.1.5
+
+## 0.2.3
+
+### Patch Changes
+
+- Force bump
+- Updated dependencies
+  - @agentflow/tools@0.1.4
+  - @agentflow/core@0.2.3
+  - create-agentflow@0.1.4
+
+## 0.2.2
+
+### Patch Changes
+
+- Force bump
+- Updated dependencies
+  - @agentflow/core@0.2.2
+  - create-agentflow@0.1.3
+  - @agentflow/tools@0.1.3
+
+## 0.2.1
+
+### Patch Changes
+
+- Force patch
+- Updated dependencies
+  - @agentflow/core@0.2.1
+  - create-agentflow@0.1.2
+  - @agentflow/tools@0.1.2
+
 ## 0.2.0
 
 ### Minor Changes