+++ packages/tools/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @agentflow/tools
 
+## 0.1.9
+
+### Patch Changes
+
+- Bymp
+- Updated dependencies
+  - @agentflow/core@0.2.8
+
+## 0.1.8
+
+### Patch Changes
+
+- Force bump
+- Updated dependencies
+  - @agentflow/core@0.2.7
+
+## 0.1.7
+
+### Patch Changes
+
+- Force bump
+- Updated dependencies
+  - @agentflow/core@0.2.6
+
+## 0.1.6
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @agentflow/core@0.2.5
+
+## 0.1.5
+
+### Patch Changes
+
+- Force
+- Updated dependencies
+  - @agentflow/core@0.2.4
+
+## 0.1.4
+
+### Patch Changes
+
+- Force bump
+- Updated dependencies
+  - @agentflow/core@0.2.3
+
+## 0.1.3
+
+### Patch Changes
+
+- Force bump
+- Updated dependencies
+  - @agentflow/core@0.2.2
+
+## 0.1.2
+
+### Patch Changes
+
+- Force patch
+- Updated dependencies
+  - @agentflow/core@0.2.1
+
 ## 0.1.1
 
 ### Patch Changes